HOCS-2611 - Remove use of healthy file
