+++ src/main/resources/application.properties
@@ -13,7 +13,6 @@ case.creator.case-service=http://localhost:8082
 case.creator.rest-client.retries=10
 case.creator.rest-client.delay=1000
 case.creator.basicauth=UNSET
-case.creator.health-file=./healthy
 case.creator.shutdown-delay-seconds=20
 case.creator.sqs.access-key=12345
 case.creator.sqs.secret-key=12345
